simplify tracing packages
